+++ vach-cli/src/commands/keypair.rs
@@ -1,4 +1,4 @@
-use vach::utils::gen_keypair;
+use vach::crypto_utils::gen_keypair;
 
 use crate::utils;
 use crate::keys::key_names;
