+++ vach/src/crypto_utils/mod.rs
@@ -1,11 +1,14 @@
-#[cfg(feature = "crypto")]
-use rand::rngs::OsRng;
-use std::{io::Read};
+#![cfg(feature = "crypto")]
+#![cfg_attr(docsrs, doc(cfg(feature = "crypto")))]
 
-use crate::global::{error::InternalError, result::InternalResult};
-
-#[cfg(feature = "crypto")]
-use crate::crypto;
+use {
+	rand::rngs::OsRng,
+	crate::{
+		crypto,
+		global::{error::InternalError, result::InternalResult},
+	},
+	std::{io::Read},
+};
 
 // A favour
 #[cfg(feature = "compression")]
@@ -13,7 +16,6 @@ pub use super::global::compressor::Compressor;
 
 /// Use this function to easily generate a [Keypair](https://docs.rs/ed25519-dalek/latest/ed25519_dalek/struct.Keypair.html) using `OsRng`
 #[inline(always)]
-#[cfg(feature = "crypto")]
 pub fn gen_keypair() -> crypto::Keypair {
 	crypto::Keypair::generate(&mut OsRng)
 }
@@ -21,8 +23,6 @@ pub fn gen_keypair() -> crypto::Keypair {
 /// Use this to read and parse a `Keypair` from a read stream
 /// ### Errors
 ///  - If the data can't be parsed into a keypair
-#[cfg(feature = "crypto")]
-#[cfg_attr(docsrs, doc(cfg(feature = "crypto")))]
 pub fn read_keypair<R: Read>(mut handle: R) -> InternalResult<crypto::Keypair> {
 	let mut keypair_bytes = [0; crate::KEYPAIR_LENGTH];
 	handle.read_exact(&mut keypair_bytes)?;
@@ -38,8 +38,6 @@ pub fn read_keypair<R: Read>(mut handle: R) -> InternalResult<crypto::Keypair> {
 /// ### Errors
 ///  - If parsing of the public key fails
 ///  - `io` errors
-#[cfg(feature = "crypto")]
-#[cfg_attr(docsrs, doc(cfg(feature = "crypto")))]
 pub fn read_public_key<T: Read>(mut handle: T) -> InternalResult<crypto::PublicKey> {
 	let mut keypair_bytes = [0; crate::PUBLIC_KEY_LENGTH];
 
@@ -55,8 +53,6 @@ pub fn read_public_key<T: Read>(mut handle: T) -> InternalResult<crypto::PublicK
 /// ### Errors
 ///  - If parsing of the secret key fails
 ///  - `io` errors
-#[cfg(feature = "crypto")]
-#[cfg_attr(docsrs, doc(cfg(feature = "crypto")))]
 pub fn read_secret_key<T: Read>(mut handle: T) -> InternalResult<crypto::SecretKey> {
 	let mut secret_bytes = [0; crate::SECRET_KEY_LENGTH];
